Sistema de upload de Fotos para usuario implementado

+++ database/migrations/2021_06_19_151012_create_insights_table.php
@@ -15,7 +15,7 @@ class CreateInsightsTable extends Migration
     {
         Schema::create('insights', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->integer('insight_id')->default(0);
+            $table->unsignedBigInteger('insight_id')->default(0);
             $table->float('faturamento');
             $table->string('tipo');
             $table->string('nome');
